app/fairy-backend/pyproject.toml
[project]
name = "fairy-backend"
version = "0.1.0"
description = "fairy-backend || FastAPI, discord.py"
requires-python = ">=3.11"
dependencies = [
    "uvicorn>=0.24.0",
    "python-dotenv>=1.0.0",
    "discord.py>=2.3.2",
    "google-genai>=0.8.3",
    "pydantic>=2.5.0",
    "fastapi[standard]>=0.120.4",
]
